[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.41 ms, Memory: 73.7 MB -BaekjoonHub

+++ 프로그래머스/lv1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -5,7 +5,7 @@ class Solution {
         ArrayList<Long> answer = new ArrayList<>();
         int count = 0 ;
         
-        for(Long i =(long)x;count<n;i+=x){
+        for(long i =x;count<n;i+=x){
             answer.add(i);
             count++;
         }
